Resturcture runtimne, offline tools and build system
We now distinguish between the project level and the roest systems level. The
project is that which contains all the assets, scripts and the build
system, it also has it's own version history. This git repo contains the
roest systems which provide the source code for runtime systems and
offline tools. It also manages its own dependencies through Cargo. It
lives in the project as a git submodule.